server/requirements.txt
flask==2.0.3
flask-sse==1.0.0
gunicorn==20.1.0
gevent==21.12.0
requests==2.27.1
aiohttp==3.8.1
celery==5.2.3
python-slugify>=7.0.0
